use keras-vis now
