+++ donkeycar/management/base.py
@@ -195,7 +195,11 @@ class CalibrateCar(BaseCommand):
                 print("Oops, {}".format(ex))
 
 
-class MakeMovie(BaseCommand):
+class MakeMovieShell(BaseCommand):
+    '''
+    take the make movie args and then call make movie command
+    with lazy imports
+    '''
     def __init__(self):
         self.deg_to_rad = math.pi / 180.0
 
@@ -218,366 +222,12 @@ class MakeMovie(BaseCommand):
         Load the images from a tub and create a movie from them.
         Movie
         '''
-        import moviepy.editor as mpy
-
         args, parser = self.parse_args(args)
 
-        if args.tub is None:
-            print("ERR>> --tub argument missing.")
-            parser.print_help()
-            return
+        from donkeycar.management.makemovie import MakeMovie
 
-        if args.type is None and args.model is not None:
-            print("ERR>> --type argument missing. Required when providing a model.")
-            parser.print_help()
-            return
-
-        if args.salient:
-            if args.model is None:
-                print("ERR>> salient visualization requires a model. Pass with the --model arg.")
-                parser.print_help()
-                return
-
-            # imported like this, we make TF conditional on use of --salient
-            # and we keep the context maintained throughout our callbacks to
-            # compute the salient mask
-            from tensorflow.python.keras import backend as K
-            import tensorflow as tf
-            os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'  
-
-        conf = os.path.expanduser(args.config)
-
-        if not os.path.exists(conf):
-            print("No config file at location: %s. Add --config to specify\
-                 location or run from dir containing config.py." % conf)
-            return
-
-        self.cfg = dk.load_config(conf)
-        self.tub = Tub(args.tub)
-        self.index = self.tub.get_index(shuffled=False)
-        start = args.start
-        self.end = args.end if args.end != -1 else len(self.index)
-        if self.end >= len(self.index):
-            self.end = len(self.index) - 1
-        num_frames = self.end - start
-        self.iRec = start
-        self.scale = args.scale
-        self.keras_part = None
-        self.convolution_part = None
-        if args.model is not None:
-            self.keras_part = get_model_by_type(args.type, cfg=self.cfg)
-            self.keras_part.load(args.model)
-            self.keras_part.compile()
-            if args.salient:
-                self.init_salient(self.keras_part.model)
-
-                #This method nested in this way to take the conditional import of TF
-                #in a manner that extends to this callback. Done this way, we avoid
-                #importing in the below method, which triggers a new cuda device allocation
-                #each call.
-                def compute_visualisation_mask(img):
-                    #from https://github.com/ermolenkodev/keras-salient-object-visualisation
-
-                    activations = self.functor([np.array([img])])
-                    activations = [np.reshape(img, (1, img.shape[0], img.shape[1], img.shape[2]))] + activations
-                    upscaled_activation = np.ones((3, 6))
-                    for layer in [5, 4, 3, 2, 1]:
-                        averaged_activation = np.mean(activations[layer], axis=3).squeeze(axis=0) * upscaled_activation
-                        output_shape = (activations[layer - 1].shape[1], activations[layer - 1].shape[2])
-                        x = tf.constant(
-                            np.reshape(averaged_activation, (1,averaged_activation.shape[0],averaged_activation.shape[1],1)),
-                            tf.float32
-                        )
-                        conv = tf.nn.conv2d_transpose(
-                            x, self.layers_kernels[layer],
-                            output_shape=(1,output_shape[0],output_shape[1], 1),
-                            strides=self.layers_strides[layer],
-                            padding='VALID'
-                        )
-                        with tf.Session() as session:
-                            result = session.run(conv)
-                        upscaled_activation = np.reshape(result, output_shape)
-                    final_visualisation_mask = upscaled_activation
-                    return (final_visualisation_mask - np.min(final_visualisation_mask))/(np.max(final_visualisation_mask) - np.min(final_visualisation_mask))
-                self.compute_visualisation_mask = compute_visualisation_mask
-
-        print('making movie', args.out, 'from', num_frames, 'images')
-        clip = mpy.VideoClip(self.make_frame,
-                             duration=((num_frames - 1) / self.cfg.DRIVE_LOOP_HZ))
-        clip.write_videofile(args.out, fps=self.cfg.DRIVE_LOOP_HZ)
-
-    def draw_user_input(self, record, img):
-        '''
-        Draw the user input as a green line on the image
-        '''
-
-        import cv2
-
-        user_angle = float(record["user/angle"])
-        user_throttle = float(record["user/throttle"])
-
-        length = self.cfg.IMAGE_H
-        a1 = user_angle * 45.0
-        l1 = user_throttle * length
-
-        mid = self.cfg.IMAGE_W // 2 - 1
-
-        p1 = tuple((mid - 2, self.cfg.IMAGE_H - 1))
-        p11 = tuple((int(p1[0] + l1 * math.cos((a1 + 270.0) * self.deg_to_rad)),
-                     int(p1[1] + l1 * math.sin((a1 + 270.0) * self.deg_to_rad))))
-
-        # user is green, pilot is blue
-        cv2.line(img, p1, p11, (0, 255, 0), 2)
-        
-    def draw_model_prediction(self, record, img):
-        '''
-        query the model for it's prediction, draw the predictions
-        as a blue line on the image
-        '''
-        if self.keras_part is None:
-            return
-
-        import cv2
-         
-        expected = self.keras_part.model.inputs[0].shape[1:]
-        actual = img.shape
-        pred_img = img
-
-        # check input depth
-        if expected[2] == 1 and actual[2] == 3:
-            pred_img = rgb2gray(pred_img)
-            pred_img = pred_img.reshape(pred_img.shape + (1,))
-            actual = pred_img.shape
-
-        if expected != actual:
-            print("expected input dim", expected, "didn't match actual dim", actual)
-            return
-
-        pilot_angle, pilot_throttle = self.keras_part.run(pred_img)
-
-        length = self.cfg.IMAGE_H
-        a2 = pilot_angle * 45.0
-        l2 = pilot_throttle * length
-
-        mid = self.cfg.IMAGE_W // 2 - 1
-
-        p2 = tuple((mid + 2, self.cfg.IMAGE_H - 1))
-        p22 = tuple((int(p2[0] + l2 * math.cos((a2 + 270.0) * self.deg_to_rad)),
-                     int(p2[1] + l2 * math.sin((a2 + 270.0) * self.deg_to_rad))))
-
-        # user is green, pilot is blue
-        cv2.line(img, p2, p22, (0, 0, 255), 2)
-
-    def draw_steering_distribution(self, record, img):
-        '''
-        query the model for it's prediction, draw the distribution of steering choices
-        '''
-        from donkeycar.parts.keras import KerasCategorical
-
-        if self.keras_part is None or type(self.keras_part) is not KerasCategorical:
-            return
-
-        import cv2
-
-        pred_img = img.reshape((1,) + img.shape)
-        angle_binned, throttle = self.keras_part.model.predict(pred_img)
-
-        x = 4
-        dx = 4
-        y = 120 - 4
-        iArgMax = np.argmax(angle_binned)
-        for i in range(15):
-            p1 = (x, y)
-            p2 = (x, y - int(angle_binned[0][i] * 100.0))
-            if i == iArgMax:
-                cv2.line(img, p1, p2, (255, 0, 0), 2)
-            else:
-                cv2.line(img, p1, p2, (200, 200, 200), 2)
-            x += dx
-
-    def init_salient(self, model):
-        #from https://github.com/ermolenkodev/keras-salient-object-visualisation
-        from tensorflow.python.keras.layers import Input, Dense, merge
-        from tensorflow.python.keras.models import Model
-        from tensorflow.python.keras.layers import Convolution2D, MaxPooling2D, Reshape, BatchNormalization
-        from tensorflow.python.keras.layers import Activation, Dropout, Flatten, Dense
-
-        input_shape = model.inputs[0].shape
-
-        img_in = Input(shape=(input_shape[1], input_shape[2], input_shape[3]), name='img_in')
-        x = img_in
-        x = Convolution2D(24, (5,5), strides=(2,2), activation='relu', name='conv1')(x)
-        x = Convolution2D(32, (5,5), strides=(2,2), activation='relu', name='conv2')(x)
-        x = Convolution2D(64, (5,5), strides=(2,2), activation='relu', name='conv3')(x)
-        x = Convolution2D(64, (3,3), strides=(2,2), activation='relu', name='conv4')(x)
-        conv_5 = Convolution2D(64, (3,3), strides=(1,1), activation='relu', name='conv5')(x)
-        self.convolution_part = Model(inputs=[img_in], outputs=[conv_5])
-
-        for layer_num in ('1', '2', '3', '4', '5'):
-            try:
-                self.convolution_part.get_layer('conv' + layer_num).set_weights(model.get_layer('conv2d_' + layer_num).get_weights())
-            except Exception as e:
-                print(e)
-                print("Failed to load layer weights for layer", layer_num)
-                raise Exception("Failed to load weights")
-
-        from tensorflow.python.keras import backend as K
-        import tensorflow as tf
-        os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
-        
-        self.inp = self.convolution_part.input                                           # input placeholder
-        self.outputs = [layer.output for layer in self.convolution_part.layers[1:]]          # all layer outputs
-        self.functor = K.function([self.inp], self.outputs)
-
-        kernel_3x3 = tf.constant(np.array([
-            [[[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]]]
-        ]), tf.float32)
-
-        kernel_5x5 = tf.constant(np.array([
-            [[[1]], [[1]], [[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]], [[1]], [[1]]],
-            [[[1]], [[1]], [[1]], [[1]], [[1]]]
-        ]), tf.float32)
-
-        self.layers_kernels = {5: kernel_3x3,
-                               4: kernel_3x3,
-                               3: kernel_5x5,
-                               2: kernel_5x5,
-                               1: kernel_5x5}
-
-        self.layers_strides = {5: [1, 1, 1, 1],
-                               4: [1, 2, 2, 1],
-                               3: [1, 2, 2, 1],
-                               2: [1, 2, 2, 1],
-                               1: [1, 2, 2, 1]}
-
-    def draw_salient(self, img):
-        # from https://github.com/ermolenkodev/keras-salient-object-visualisation
-        import cv2
-        alpha = 0.004
-        beta = 1.0 - alpha
-
-        expected = self.keras_part.model.inputs[0].shape[1:]
-        actual = img.shape
-        pred_img = img
-
-        # check input depth
-        if expected[2] == 1 and actual[2] == 3:
-            pred_img = rgb2gray(pred_img)
-            pred_img = pred_img.reshape(pred_img.shape + (1,))
-            actual = pred_img.shape
-
-        salient_mask = self.compute_visualisation_mask(pred_img)
-        salient_mask_stacked = np.dstack((salient_mask, salient_mask))
-        salient_mask_stacked = np.dstack((salient_mask_stacked, salient_mask))
-        blend = cv2.addWeighted(img.astype('float32'), alpha, salient_mask_stacked, beta, 0.0)
-        return blend
-
-    def make_frame(self, t):
-        '''
-        Callback to return an image from from our tub records.
-        This is called from the VideoClip as it references a time.
-        We don't use t to reference the frame, but instead increment
-        a frame counter. This assumes sequential access.
-        '''
-
-        if self.iRec >= self.end or self.iRec >= len(self.index):
-            return None
-
-        rec_ix = self.index[self.iRec]
-        rec = self.tub.get_record(rec_ix)
-        image = rec['cam/image_array']
-
-        if self.convolution_part:
-            image = self.draw_salient(image)
-            image = image * 255
-            image = image.astype('uint8')
-        
-        self.draw_user_input(rec, image)
-        if self.keras_part is not None:
-            self.draw_model_prediction(rec, image)
-            self.draw_steering_distribution(rec, image)
-
-        if self.scale != 1:
-            import cv2
-            h, w, d = image.shape
-            dsize = (w * self.scale, h * self.scale)
-            image = cv2.resize(image, dsize=dsize, interpolation=cv2.INTER_CUBIC)
-
-        self.iRec += 1
-        # returns a 8-bit RGB array
-        return image
-
-
-class Sim(BaseCommand):
-    '''
-    Start a websocket SocketIO server to talk to a donkey simulator    
-    '''
-    
-    def parse_args(self, args):
-        parser = argparse.ArgumentParser(prog='sim')
-        parser.add_argument('--model', help='the model to use for predictions')
-        parser.add_argument('--config', default='./config.py', help='location of config file to use. default: ./config.py')
-        parser.add_argument('--type', default='categorical', help='model type to use when loading. categorical|linear')
-        parser.add_argument('--top_speed', default='3', help='what is top speed to drive')
-        parsed_args = parser.parse_args(args)
-        return parsed_args, parser
-
-    def run(self, args):
-        '''
-        Start a websocket SocketIO server to talk to a donkey simulator
-        '''
-        import socketio
-        from donkeycar.parts.simulation import SteeringServer
-        from donkeycar.parts.keras import KerasCategorical, KerasLinear
-
-        args, parser = self.parse_args(args)
-
-        cfg = load_config(args.config)
-
-        if cfg is None:
-            return
-
-        #TODO: this logic should be in a pilot or model handler part.
-        if args.type == "categorical":
-            kl = KerasCategorical()
-        elif args.type == "linear":
-            kl = KerasLinear(num_outputs=2)
-        else:
-            print("ERR>> Didn't recognize type:", args.type)
-            parser.print_help()
-            return
-
-        #can provide an optional image filter part
-        img_stack = None
-
-        #load keras model
-        kl.load(args.model)  
-
-        #start socket server framework
-        sio = socketio.Server()
-
-        top_speed = float(args.top_speed)
-
-        #start sim server handler
-        ss = SteeringServer(sio, kpart=kl, top_speed=top_speed, image_part=img_stack)
-                
-        #register events and pass to server handlers
-
-        @sio.on('telemetry')
-        def telemetry(sid, data):
-            ss.telemetry(sid, data)
-
-        @sio.on('connect')
-        def connect(sid, environ):
-            ss.connect(sid, environ)
-
-        ss.go(('0.0.0.0', 9090))
-
+        mm = MakeMovie()
+        mm.run(args, parser)
 
 
 class TubCheck(BaseCommand):
@@ -716,7 +366,7 @@ class ShowCnnActivations(BaseCommand):
         import matplotlib.pyplot as plt
         self.plt = plt
 
-    def get_activations(self, image_path, model_path):
+    def get_activations(self, image_path, model_path, cfg):
         '''
         Extracts features from an image
 
@@ -728,7 +378,7 @@ class ShowCnnActivations(BaseCommand):
         image_path = os.path.expanduser(image_path)
 
         model = load_model(model_path)
-        image = self.plt.imread(image_path)[None, ...]
+        image = load_scaled_image_arr(image_path, cfg)[None, ...]
 
         conv_layer_names = self.get_conv_layers(model)
         input_layer = model.get_layer(name='img_in').input
@@ -770,13 +420,15 @@ class ShowCnnActivations(BaseCommand):
         parser = argparse.ArgumentParser(prog='cnnactivations', usage='%(prog)s [options]')
         parser.add_argument('--image', help='path to image')
         parser.add_argument('--model', default=None, help='path to model')
+        parser.add_argument('--config', default='./config.py', help='location of config file to use. default: ./config.py')
         
         parsed_args = parser.parse_args(args)
         return parsed_args
 
     def run(self, args):
         args = self.parse_args(args)
-        activations = self.get_activations(args.image, args.model)
+        cfg = load_config(args.config)
+        activations = self.get_activations(args.image, args.model, cfg)
         self.create_figure(activations)
 
 
@@ -870,8 +522,7 @@ def execute_from_command_line():
             'tubhist': ShowHistogram,
             'tubplot': ShowPredictionPlots,
             'tubcheck': TubCheck,
-            'makemovie': MakeMovie,            
-            'sim': Sim,
+            'makemovie': MakeMovieShell,            
             'createjs': CreateJoystick,
             'consync': ConSync,
             'contrain': ConTrain,
